@@ -1,32 +1,34 @@
 apply plugin: 'com.android.application'
 
+apply plugin: 'embrace-swazzler'
+
 apply plugin: 'kotlin-android'
 
-apply plugin: 'kotlin-android-extensions'
+apply plugin: 'kotlin-parcelize'
 
 apply plugin: 'com.google.gms.google-services'
 
 android {
-    compileSdkVersion 30
+    compileSdkVersion 33
     buildToolsVersion "30.0.3"
     def keys = rootProject.extensions.getByName("ext")
 
-    apply from: rootProject.file('app/release.gradle')
+    //apply from: rootProject.file('app/release.gradle')
 
     defaultConfig {
         applicationId "com.nelsito.travelplan"
         minSdkVersion 25
-        targetSdkVersion 30
+        targetSdkVersion 33
         versionCode 1
         versionName "1.0"
-        resValue "string", "api_key", keys["api_key"]
-        resValue "string", "send_grid_api_key", keys["send_grid"]
+        //resValue "string", "api_key", keys["api_key"]
+        //resValue "string", "send_grid_api_key", keys["send_grid"]
 //(project.findProperty("PLACES_API_KEY") ?: "")
         testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
         release {
-            signingConfig signingConfigs.release
+//            signingConfig signingConfigs.release
             minifyEnabled false
             proguardFiles getDefaultProguardFile('proguard-android-optimize.txt'), 'proguard-rules.pro'
         }
