@@ -8,52 +8,45 @@ import androidx.recyclerview.widget.ListAdapter
 import androidx.recyclerview.widget.RecyclerView
 import com.bumptech.glide.Glide
 import com.nelsito.travelplan.R
-import kotlinx.android.synthetic.main.user_list_item.view.*
+import com.nelsito.travelplan.databinding.TripListItemBinding
+import com.nelsito.travelplan.databinding.UserListItemBinding
 
 class UserListAdapter(private val context: Context, private val clickListener: (UserListItem) -> Unit): ListAdapter<UserListItem, RecyclerView.ViewHolder>(UserDiffCallback()) {
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): RecyclerView.ViewHolder {
-        val inflater = LayoutInflater.from(parent.context)
-        return UserViewHolder(
-            inflater.inflate(
-                R.layout.user_list_item,
-                parent,
-                false
-            )
-        )
+        val bindingList = UserListItemBinding.inflate(LayoutInflater.from(parent.context),parent,false)
+        return UserViewHolder(bindingList)
     }
 
     override fun onBindViewHolder(holder: RecyclerView.ViewHolder, position: Int) {
         (holder as UserViewHolder).bind(getItem(position) as UserListItem, context, clickListener)
     }
 
-    private class UserViewHolder(
-        itemView: View
-    ) : RecyclerView.ViewHolder(itemView) {
+    private class UserViewHolder(private val binding: UserListItemBinding) : RecyclerView.ViewHolder(binding.root) {
         fun bind(userListItem: UserListItem, context: Context,
                  clickListener: (UserListItem) -> Unit) {
-            itemView.txt_username.text = userListItem.username
-            itemView.txt_email.text = userListItem.email
+            binding.txtUsername.text = userListItem.username
+            binding.txtEmail.text = userListItem.email
 
             if (userListItem.disabled) {
-                itemView.lbl_manager.visibility = View.GONE
-                itemView.lbl_admin.visibility = View.GONE
-                itemView.lbl_disabled.visibility = View.VISIBLE
+                binding.lblManager.visibility = View.GONE
+                binding.lblAdmin.visibility = View.GONE
+                binding.lblDisabled.visibility = View.VISIBLE
             }
             when(userListItem.role) {
                 "admin" -> {
-                    itemView.lbl_disabled.visibility = View.GONE
-                    itemView.lbl_manager.visibility = View.GONE
-                    itemView.lbl_admin.visibility = View.VISIBLE
+                    binding.lblDisabled.visibility = View.GONE
+                    binding.lblManager.visibility = View.GONE
+                    binding.lblAdmin.visibility = View.VISIBLE
                 }
                 "manager" -> {
-                    itemView.lbl_disabled.visibility = View.GONE
-                    itemView.lbl_admin.visibility = View.GONE
-                    itemView.lbl_manager.visibility = View.VISIBLE
+                    binding.lblDisabled.visibility = View.GONE
+                    binding.lblAdmin.visibility = View.GONE
+                    binding.lblManager.visibility = View.VISIBLE
                 }
                 else -> {
-                    itemView.lbl_disabled.visibility = View.GONE
-                    itemView.lbl_manager.visibility = View.GONE
-                    itemView.lbl_admin.visibility = View.GONE
+                    binding.lblDisabled.visibility = View.GONE
+                    binding.lblManager.visibility = View.GONE
+                    binding.lblAdmin.visibility = View.GONE
                 }
             }
 
@@ -61,9 +54,9 @@ class UserListAdapter(private val context: Context, private val clickListener: (
                 .load(userListItem.photoUrl)
                 .centerCrop()
                 .placeholder(context.getDrawable(R.drawable.ic_person_white_24dp))
-                .into(itemView.img_avatar)
+                .into(binding.imgAvatar)
 
-            itemView.setOnClickListener { clickListener(userListItem) }
+            binding.root.setOnClickListener { clickListener(userListItem) }
         }
     }
 }
